pyproject.toml
[project]
name = "graphql-server"
version = "0.1.0"
description = "A Server using FastAPI and Strawberry GraphQL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "sqlalchemy>=2.0.43",
    "strawberry-graphql[debug-server]>=0.282.0",
]
